Pp delete book

+++ LibSys2.0/LibSys2.0/ViewModels/LibrarianViewModel.cs
@@ -22,6 +22,8 @@ namespace LibrarySystem.ViewModels
         public Book SelectedBook { get; set; }
         public eBook SelectedeBook { get; set; }
 
+        public string ReasonToDelete { get; set; }
+
         public List<Event> ListOfEvents { get; set; }
         public ObservableCollection<Book> Books { get; set; }
         public ObservableCollection<eBook> eBooks { get; set; }
@@ -37,6 +39,8 @@ namespace LibrarySystem.ViewModels
         public ReactiveCommand<Unit, Unit> AddeBookCommand { get; set; }
         public ReactiveCommand<Unit, Unit> AddEventCommand { get; set; }
         public ReactiveCommand<object, Unit> ToggleHidden { get; set; }
+        public ReactiveCommand<object, Unit> ToggleVisible { get; set; }
+
 
         #endregion
         public LibrarianViewModel()
@@ -52,7 +56,8 @@ namespace LibrarySystem.ViewModels
             RemoveBookCommand = ReactiveCommand.CreateFromTask((int id) => RemoveBookCommandMethod(id));
             AddeBookCommand = ReactiveCommand.CreateFromTask(() => AddeBookCommandMethod());
             AddEventCommand = ReactiveCommand.CreateFromTask(() => AddEventCommandMethod());
-            ToggleHidden = ReactiveCommand.CreateFromTask((object param) => ToggleHiddenCommandMethod(param));
+            ToggleHidden = ReactiveCommand.CreateFromTask((object param) => HiddenCommandMethod(param));
+            ToggleVisible = ReactiveCommand.CreateFromTask((object param) => VisibleCommandMethod(param));
             LoadDataAsync();
         }
 
@@ -89,7 +94,7 @@ namespace LibrarySystem.ViewModels
             await ClearBookLines("books");
 
         }
-
+        #endregion
         public async Task UpdateBookCommandMethod(Book book)
         #region ...
         {
@@ -100,9 +105,15 @@ namespace LibrarySystem.ViewModels
         public async Task RemoveBookCommandMethod(int id)
         #region ...
         {
+            if(ReasonToDelete == null)
+            {
+                MessageBox.Show("Fyll i anledning!");
+                return;
+            }
             await bookRepo.Delete(id);
             await LoadBooks();
         }
+        #endregion
         /// <summary>
         /// Checks and adds a E-Book to the database
         /// </summary>
@@ -139,19 +150,35 @@ namespace LibrarySystem.ViewModels
             await LoadEbooks();
             await ClearBookLines("ebooks");
         }
+        #endregion
         public async Task AddEventCommandMethod()
         #region ...
         {
 
         }
-        #endregion
-        public async Task ToggleHiddenCommandMethod(object arg)
-        #region ...
+
+
+        public async Task VisibleCommandMethod(object arg)
         {
             var button = (Button)arg;
-            button.IsEnabled = button.IsEnabled ? false : true;
+            button.IsEnabled = true;
+            ReasonToDelete = "";
+            this.OnPropertyChanged(nameof(ReasonToDelete));
         }
+        public async Task HiddenCommandMethod(object arg)
+        {
+            var button = (Button)arg;
+            button.IsEnabled = false;
+        }
+
+
         #endregion
+        //public async Task ToggleHiddenCommandMethod(object arg)
+        
+        //{
+        //    var button = (Button)arg;
+        //    button.IsEnabled = button.IsEnabled ? false : true;
+        //}
 
         public async void LoadDataAsync()
         {
@@ -230,5 +257,6 @@ namespace LibrarySystem.ViewModels
         //        events = new ObservableCollection<Event>(eventIE.ToList());
         //    }
         //}
+        
     }
 }
